# docs/swagger.yaml
definitions:
  main.ProductCreateRequest:
    properties:
      category_id:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      reference:
        type: string
      status:
        type: string
      stock_city:
        type: string
      supplier_id:
        type: string
    required:
    - name
    - reference
    type: object
  main.ProductUpdateRequest:
    properties:
      category_id:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      reference:
        type: string
      status:
        type: string
      stock_city:
        type: string
      supplier_id:
        type: string
    type: object
  models.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
info:
  contact: {}
paths:
  /api/statistics/products-per-category:
    get:
      description: 'Add "Authorization: Bearer {your_token}" in headers to authenticate'
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Statistics products per category
  /api/statistics/products-per-supplier:
    get:
      description: 'Add "Authorization: Bearer {your_token}" in headers to authenticate'
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Statistics products per supplier
  /distance:
    get:
      description: |-
        Add "Authorization: Bearer {your_token}" in headers to authenticate
        Calculate Distance from your location to a city
      parameters:
      - description: City
        in: query
        name: city
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Calculate Distance
  /products:
    get:
      description: |-
        Fetch products with pagination and filtering
        Add "Authorization: Bearer {your_token}" in headers to authenticate
      parameters:
      - description: Number of products per page
        in: query
        name: perPage
        type: integer
      - description: Field to filter by (e.g., supplier, category)
        in: query
        name: field
        type: string
      - description: Values of field
        in: query
        name: values
        type: array
      - description: The last reference of previous page
        in: query
        name: last_reference
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Get products
    post:
      description: 'Add "Authorization: Bearer {your_token}" in headers to authenticate'
      parameters:
      - description: Product filter request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.ProductCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Create product
  /products/:id:
    delete:
      description: 'Add "Authorization: Bearer {your_token}" in headers to authenticate'
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Delete product
    put:
      description: 'Add "Authorization: Bearer {your_token}" in headers to authenticate'
      parameters:
      - description: Product filter request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.ProductUpdateRequest'
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Update product
  /products/categories:
    get:
      description: 'Add "Authorization: Bearer {your_token}" in headers to authenticate'
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Get all categories of products
  /products/cities:
    get:
      description: 'Add "Authorization: Bearer {your_token}" in headers to authenticate'
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Get all cities of products
  /products/export:
    get:
      description: 'Add "Authorization: Bearer {your_token}" in headers to authenticate'
      parameters:
      - description: Field to filter by (e.g., supplier, category)
        in: query
        name: field
        type: string
      - description: Values of field
        in: query
        name: values
        type: array
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Export products
  /products/suppliers:
    get:
      description: 'Add "Authorization: Bearer {your_token}" in headers to authenticate'
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Get all suppliers of products
  /users/sign-in:
    post:
      description: signin to get token to use api
      parameters:
      - description: Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: SignIn
  /users/sign-up:
    post:
      description: create account for user to use api
      parameters:
      - description: Create user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: SignUp
swagger: "2.0"